Devel (#14)
* merge cpp and roscpp ultisnips definition

* minor changes

* fixing bash folding

* added matej

* Added Matej's Asus configs and scripts

* put back S-arrow for tmux windows

* updated dotprofiler, added custom commenting rules

* fixing tmux movement bindings

* Added backup.sh

* renamed backup script

* adding git reset --hard hook to dotprofiler

* fixing git reset

* updated dotprofiler

* updated dotprofiler scripts

* added monitor switching for GPD

* moved gpdSetMonitor to scripts

* moved set_color_scheme to as separate scrirp

* custom brightness control

* updating keybindings

* updating keybindings

* unified Xresources naming

* fixed comment

+++ appconfig/shell/commons.sh
@@ -32,43 +32,51 @@ killp() {
   else
     pes=$1
   fi
-  
-	for child in $(ps -o pid,ppid -ax | \
-   awk "{ if ( \$2 == $pes ) { print \$1 }}")
-  do
-    # echo "Killing child process $child because ppid = $pes"
-    killp $child
-  done
-        
-  # echo "killing $1"
-  kill -9 "$1" > /dev/null 2> /dev/null
+
+  for child in $(ps -o pid,ppid -ax | \
+    awk "{ if ( \$2 == $pes ) { print \$1 }}")
+do
+  # echo "Killing child process $child because ppid = $pes"
+  killp $child
+done
+
+# echo "killing $1"
+kill -9 "$1" > /dev/null 2> /dev/null
 }
 
 killSession() {
-  echo "killing session"
-  pids=`tmux list-panes -s -F "#{pane_pid} #{pane_current_command}" | grep -v tmux | awk {'print $1'}`
 
-  for pid in $pids; do
-    killp "$pid" &
-    # echo $pid
-  done
+  if [ ! -z "$TMUX" ]; then
 
-  SESSION_NAME=`tmux display-message -p '#S'`
-  tmux kill-session -t "$SESSION_NAME"
+    echo "killing session"
+    pids=`tmux list-panes -s -F "#{pane_pid} #{pane_current_command}" | grep -v tmux | awk {'print $1'}`
+
+    for pid in $pids; do
+      killp "$pid" &
+    done
+
+    SESSION_NAME=`tmux display-message -p '#S'`
+    tmux kill-session -t "$SESSION_NAME"
+
+  else
+
+    exit
+
+  fi
 }
 alias :qa="killSession"
 
 gitPullPush() {
 
- branch=`git branch | grep \* | sed 's/\* \([a-Z]*\)/\1/'`
+  branch=`git branch | grep \* | sed 's/\* \([a-Z]*\)/\1/'`
 
- if [ $# -eq 0 ]; then
-  git pull origin "$branch"
-  git push origin "$branch"
- else
-  git pull "$1" "$branch"
-  git push "$1" "$branch"
- fi
+  if [ $# -eq 0 ]; then
+    git pull origin "$branch"
+    git push origin "$branch"
+  else
+    git pull "$1" "$branch"
+    git push "$1" "$branch"
+  fi
 }
 
 getVideoThumbnail () {
@@ -93,42 +101,40 @@ getVideoThumbnail () {
 # Other "git" features should not be changed
 git() {
 
-  case $* in
+  case $* in pull*|checkout*|"reset --hard")
 
-    pull*|checkout*)
+    # give me the path to root of the repo we are in
+    ROOT_DIR=`git rev-parse --show-toplevel` 2> /dev/null
 
-      # give me the path to root of the repo we are in
-      ROOT_DIR=`git rev-parse --show-toplevel` 2> /dev/null
+    if [[ "$?" == "0" ]]; then
 
-      if [[ "$?" == "0" ]]; then
+      # if we are in the 'linux-setup' repo, use the git profiler
+      if [[ "$ROOT_DIR" == "$GIT_PATH/linux-setup" ]]; then
 
-        # if we are in the 'linux-setup' repo, use the git profiler
-        if [[ "$ROOT_DIR" == "$GIT_PATH/linux-setup" ]]; then
+        PROFILER="$GIT_PATH/linux-setup/submodules/dotprofiler/profiler.sh"
 
-          PROFILER="$GIT_PATH/linux-setup/submodules/dotprofiler/profiler.sh"
+        bash -c "$PROFILER backup $GIT_PATH/linux-setup/appconfig/dotprofiler/file_list.txt"
 
-          bash -c "$PROFILER backup $GIT_PATH/linux-setup/appconfig/dotprofiler/file_list.txt"
+        command git "$@"
 
-          command git "$@"
+        if [[ "$?" == "0" ]]; then
 
-          if [[ "$?" == "0" ]]; then
+          bash -c "$PROFILER deploy $GIT_PATH/linux-setup/appconfig/dotprofiler/file_list.txt"
 
-            bash -c "$PROFILER deploy $GIT_PATH/linux-setup/appconfig/dotprofiler/file_list.txt"
-
-          fi
-
-        else
-          command git "$@"
         fi
 
       else
         command git "$@"
       fi
 
-      ;;
-    *)
+    else
       command git "$@"
-      ;;
+    fi
+
+    ;;
+  *)
+    command git "$@"
+    ;;
 
   esac
 }
@@ -148,34 +154,6 @@ sourceShellDotfile() {
 }
 alias sb="sourceShellDotfile"
 
-setColorScheme() {
-
-  case "$SHELL" in 
-    *bash*)
-      RCFILE="$HOME/.bashrc"
-      ;;
-    *zsh*)
-      RCFILE="$HOME/.zshrc"
-      ;;
-  esac
-
-  export COLOR_SCHEME="$1"
-
-  # change the variable in bashrc
-  /usr/bin/vim -u "$GIT_PATH/linux-setup/submodules/dotprofiler/epigen/epigen.vimrc" -E -s -c "%g/.*PROFILER.*COLORSCHEME.*/norm ^/COLORSCHEMEciwCOLORSCHEME_$COLOR_SCHEME" -c "wqa" -- "$RCFILE"
-
-  source "$RCFILE"
-
-  cd "$GIT_PATH/linux-setup"
-  ./backup_and_deploy.sh
-
-  # reload configuration for urxvt
-  xrdb ~/.Xresources
-
-  # restart i3
-  i3-msg restart
-}
-
 # special code for i3 users
 if [ "$USE_I3" = "true" ]; then
 
@@ -186,15 +164,15 @@ if [ "$USE_I3" = "true" ]; then
   xset r rate 350 55
 
   echo '#!/bin/bash
-echo '"$ROS_IP" > ~/.i3/ros_ip.sh
+  echo '"$ROS_IP" > ~/.i3/ros_ip.sh
   chmod +x ~/.i3/ros_ip.sh
 
   echo '#!/bin/bash
-echo '"$ROS_MASTER_URI" | sed 's/http:\/\/\(.*\):.*/\1/' > ~/.i3/ros_master_uri.sh
+  echo '"$ROS_MASTER_URI" | sed 's/http:\/\/\(.*\):.*/\1/' > ~/.i3/ros_master_uri.sh
   chmod +x ~/.i3/ros_master_uri.sh
-  
+
   echo '#!/bin/bash
-echo '"$UAV_NAME" > ~/.i3/uav_name.sh
+  echo '"$UAV_NAME" > ~/.i3/uav_name.sh
   chmod +x ~/.i3/uav_name.sh
 
   export TERM=rxvt-unicode-256color
